requirements.txt
django==4.1.5
djangorestframework==3.14.0
pillow==9.4.0
djoser==2.1.0
flake8==6.0.0